requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiobotocore==2.5.0
    # via -r requirements.in
aiohttp==3.8.4
    # via aiobotocore
aioitertools==0.11.0
    # via aiobotocore
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via httpcore
async-generator==1.10
    # via trio
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   cattrs
    #   outcome
    #   trio
    #   trio-asyncio
beautifulsoup4==4.12.0
    # via -r requirements.in
black==23.3.0
    # via -r requirements.in
botocore==1.29.76
    # via aiobotocore
botocore-stubs==1.29.104
    # via types-aiobotocore
cattrs==22.2.0
    # via -r requirements.in
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
charset-normalizer==3.1.0
    # via aiohttp
click==8.1.3
    # via
    #   black
    #   typer
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.16.3
    # via httpx
httpx[http2]==0.23.3
    # via -r requirements.in
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   rfc3986
    #   trio
    #   yarl
isort==5.12.0
    # via -r requirements.in
jmespath==1.0.1
    # via botocore
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
outcome==1.2.0
    # via
    #   trio
    #   trio-asyncio
packaging==23.0
    # via black
pathspec==0.11.1
    # via black
platformdirs==3.2.0
    # via black
python-dateutil==2.8.2
    # via botocore
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
    #   trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.4
    # via beautifulsoup4
structlog==22.3.0
    # via -r requirements.in
tomli-w==1.0.0
    # via -r requirements.in
toolz==0.12.0
    # via -r requirements.in
trio==0.22.0
    # via
    #   -r requirements.in
    #   trio-asyncio
trio-asyncio==0.12.0
    # via -r requirements.in
typer==0.7.0
    # via -r requirements.in
types-aiobotocore[essential]==2.5.0.post2
    # via -r requirements.in
types-aiobotocore-cloudformation==2.5.0.post1
    # via types-aiobotocore
types-aiobotocore-dynamodb==2.5.0.post1
    # via types-aiobotocore
types-aiobotocore-ec2==2.5.0.post1
    # via types-aiobotocore
types-aiobotocore-lambda==2.5.0.post1
    # via types-aiobotocore
types-aiobotocore-rds==2.5.0.post1
    # via types-aiobotocore
types-aiobotocore-s3==2.5.0.post1
    # via types-aiobotocore
types-aiobotocore-sqs==2.5.0.post1
    # via types-aiobotocore
types-awscrt==0.16.13.post1
    # via botocore-stubs
typing-extensions==4.5.0
    # via
    #   types-aiobotocore-cloudformation
    #   types-aiobotocore-dynamodb
    #   types-aiobotocore-ec2
    #   types-aiobotocore-lambda
    #   types-aiobotocore-rds
    #   types-aiobotocore-s3
    #   types-aiobotocore-sqs
urllib3==1.26.15
    # via botocore
wrapt==1.15.0
    # via aiobotocore
yarl==1.8.2
    # via aiohttp
